A* working!
Still a little bit messy, still including the test code, but the A_star class seems to be working well.

TODO: handle tunnels, and when the road is going under another road. A solution of this may require the path.xml file of GTA V, and would have to redefine which pixels are roads, so it might not be feasible or very productive for this project.